Moved static IFunction methods
